sns: rename generic pwr keys with a typed prefix
Ratio, Mains are specific to a magnitude type.
Make keys use global index, and do a (temporary) WebUI workaround
to use 0 index instead of the global key.
Migrate to the new settings keys.

Local magnitude-specific index is still in play, pending changes in the rest of sensors
(ADE7953, PZEM004T{,V30}, ...) to remove it completely.

Local sensor index could still be useful though, to help out with some settings
in the future as we don't really enforce magnitude type uniqueness within sensors.

+++ code/espurna/settings.h
@@ -185,7 +185,7 @@ String settingsQueryDefaults(const String& key);
 // --------------------------------------------------------------------------
 
 void moveSetting(const String& from, const String& to);
-void moveSetting(const String& from, const String& to, unsigned int index);
+void moveSetting(const String& from, const String& to, size_t index);
 void moveSettings(const String& from, const String& to);
 
 template <typename T, typename = typename settings::internal::enable_if_not_arduino_string<T>::type>
